Persist token expirations with absolute times instead of relative.
When storing a request or access token, persist the token expiration
time with an absolute timestamp. By doing so an application can
accurately determine if a token is still valid even if it is shutdown
and restarted after the token expires.  Previously a relative timestamp
was used, which meant a token restored from the persistent store always
claimed to be valid even if it wasn't.